+++ src/PluginInfo.cs
@@ -1,9 +1,9 @@
-namespace FikaWeaponPickupFix
+﻿namespace FikaWeaponPickupFix
 {
     public static class PluginInfo
     {
         public const string GUID = "com.dematch.fikaweaponpickupfix";
         public const string NAME = "Fika Weapon Pickup Fix";
-        public const string VERSION = "1.1.0";
+        public const string VERSION = "1.5.2";
     }
 }